requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.2.*
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
Markdown==3.0.1
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.20.0
six==1.10.0
text-unidecode==1.2
urllib3==1.24.2
django-rest-swagger
